+++ src/main/java/com/youzan/nsq/client/entity/Response.java
@@ -27,6 +27,7 @@ public enum Response {
     E_BAD_TAG("E_BAD_TAG"),
     E_TAG_NOT_SUPPORT ("E_TAG_NOT_SUPPORT"),
     E_SUB_EXTEND_NEED ("E_SUB_EXTEND_NEED"),
+    E_EXT_NOT_SUPPORT ("E_EXT_NOT_SUPPORT")
     ;
 
     private final String content;